requirements.txt
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
av==12.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==42.0.8
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
entrypoints==0.4
executing==2.0.1
extra-streamlit-components==0.1.71
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.136.0
google-apps-meet==0.1.6
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-speech==2.26.0
google-crc32c==1.5.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ifaddr==0.2.0
importlib-metadata==6.11.0
ipykernel==6.29.4
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langsmith==0.1.83
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.7
orjson==3.10.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pyee==11.1.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
pylibsrtp==0.10.0
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
selenium==4.22.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.30.0
streamlit-authenticator==0.3.2
streamlit-calendar==1.2.0
streamlit-card==1.0.2
streamlit-javascript==0.1.5
streamlit-modal==0.1.2
streamlit-on-Hover-tabs==1.0.1
streamlit-option-menu==0.3.13
streamlit-pills==0.3.0
streamlit-webrtc==0.47.7
sympy==1.12.1
tenacity==8.4.2
threadpoolctl==3.5.0
tikapi==3.1.60
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.2
validators==0.28.3
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
